requirements.txt
python-dateutil >= 2.6.0
ujson >= 1.35
bintrees >= 2.0.7